feature: retries added
